feat: 实现完整MVP训练系统并完成首次训练
核心模块实现：
- 多阶段GAT模型（559K参数）
  - 局部GAT编码器（3层×8头）
  - 全局虚拟节点注意力（4头）
  - 规则聚焦注意力
  - 多路径融合机制
- 完整训练编排器
  - 训练/验证循环
  - Checkpoint管理
  - Loss监控

首次训练结果：
- 数据集: 100场景（80训练+20验证）
- Epochs: 10
- Val Loss: 3.67 → 0.95 (↓74%)
- Checkpoint: 4个文件已保存

Bug修复：
- 修复dtype不匹配（float32 vs float64）
- 修复DataLoader collate问题

设计依据: Design-ITER-2025-01.md v2.0 §3.3-3.5
清单项: 1-20 全部完成
总代码量: ~950行新增

+++ src/traffic_rules/rules/red_light.py
@@ -69,7 +69,10 @@ def compute_rule_score_differentiable(
         light_weights = light_probs[:, 0]
     
     # Step 2: 计算分段距离-速度评分 f_dv(d, v)
-    f_dv = torch.zeros(B, device=device)
+    # 统一dtype为float32
+    distances = distances.float()
+    velocities = velocities.float()
+    f_dv = torch.zeros(B, device=device, dtype=torch.float32)
     
     # 情况1：已过线（d < 0）
     # 物理意义：车辆闯过停止线，距离越远（负得越多）违规越严重
